requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
tensorflow
nltk
transformers
torch
lime
google-generativeai
matplotlib
tqdm
joblib
protobuf
requests